+++ src/Manifold/DevNull.php
@@ -14,7 +14,7 @@ use DecodeLabs\Systemic\ManifoldAbstract;
 class DevNull extends ManifoldAbstract
 {
     /**
-     * @return array<int, resource|array<string>>
+     * @return array<int,resource|array<string>>
      */
     public function getDescriptors(): array
     {