+++ src/Manifold/Pipe.php
@@ -14,7 +14,7 @@ use DecodeLabs\Systemic\ManifoldAbstract;
 class Pipe extends ManifoldAbstract
 {
     /**
-     * @return array<int, resource|array<string>>
+     * @return array<int,resource|array<string>>
      */
     public function getDescriptors(): array
     {