+++ src/ProcessTrait.php
@@ -12,6 +12,9 @@ namespace DecodeLabs\Systemic;
 use DecodeLabs\Exceptional;
 use Throwable;
 
+/**
+ * @phpstan-require-implements Process
+ */
 trait ProcessTrait
 {
     protected int $processId;
@@ -72,7 +75,7 @@ trait ProcessTrait
                 $write = false;
             } elseif (self::isProcessIdLive($oldPid)) {
                 throw Exceptional::Runtime(
-                    'PID file ' . basename($path) . ' already exists and is live with pid of ' . $oldPid
+                    message: 'PID file ' . basename($path) . ' already exists and is live with pid of ' . $oldPid
                 );
             }
         }
@@ -83,10 +86,8 @@ trait ProcessTrait
                 file_put_contents($path, $pid);
             } catch (Throwable $e) {
                 throw Exceptional::Runtime(
-                    'Unable to write PID file',
-                    [
-                        'previous' => $e
-                    ]
+                    message: 'Unable to write PID file',
+                    previous: $e
                 );
             }
         }